+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/HibernatePresetLookupDao.java
@@ -45,7 +45,7 @@ public class HibernatePresetLookupDao<E extends Entity<?>, PE extends Bean> impl
         try {
             DetachedCriteria criteria = DetachedCriteria.forClass(classPE);
             presetCriteriaMaker.makeCriteria(criteria, preset, objs);
-            //noinspection unchecked
+            @SuppressWarnings("unchecked")
             List<PE> byCriteria = (List<PE>) template.findByCriteria(criteria);
             return byCriteria.stream().map(entityBeanTransformer::reverseTransform).collect(Collectors.toList());
         } catch (Exception e) {
@@ -58,7 +58,7 @@ public class HibernatePresetLookupDao<E extends Entity<?>, PE extends Bean> impl
         try {
             DetachedCriteria criteria = DetachedCriteria.forClass(classPE);
             presetCriteriaMaker.makeCriteria(criteria, preset, objs);
-            //noinspection unchecked
+            @SuppressWarnings("unchecked")
             List<PE> byCriteria = (List<PE>) template.findByCriteria(criteria,
                     pagingInfo.getPage() * pagingInfo.getRows(), pagingInfo.getRows());
             return byCriteria.stream().map(entityBeanTransformer::reverseTransform).collect(Collectors.toList());