+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/dao/MemorySingleObjectDao.java
@@ -44,7 +44,7 @@ public class MemorySingleObjectDao<K extends Key, E extends Entity<K>> implement
     @Override
     public void put(E entity) throws DaoException {
         try {
-            //noinspection unchecked
+            @SuppressWarnings("unchecked")
             E newEntity = (E) BeanUtils.cloneBean(entity);
             newEntity.setKey(key);
             memory.set(newEntity);